Restructure as library.
More useful to produce a library which is linked in to a game project
rather than copying this repo every time. Makes upstream changes easier
to pull in to game, and removes temptation to change local copy of
engine instead of updating the engine project directly.

However, this reducing the ability to hack the engine - only allowing a
user to sub class to change behaviour of the engine. I think this is
okay - there is plenty of hackability in this approach and if they
really want to the user can still copy over the source and edit
directly.